refactor(cli): extract condensed events from operator_cli
Keep core/src/operator_cli.rs at the v0.36.24 line budget by moving
the new events command into core/src/condensed_events.rs. poll-events
stays unchanged in operator_cli.
